+++ Fossil/src/Game.cpp
@@ -4,8 +4,12 @@
 
 int main() {
 	//	Log message noting that the game started
-	logMessage("Game started", "game.txt", true);
+	logMessage(". . . Starting . . . \n", "gamelogs.txt", true);
+
+	//	Create an array to store the settings and get settings from gameprefs
+	std::vector<std::string> displaySettings = getSettings("Display");
+	std::vector<std::string> graphicsSettings = getSettings("Graphics");
 
 	//	Log message noting that the game has exited
-	logMessage("Game finished", "game.txt", false);
+	logMessage("\n . . . Ending . . .", "gamelogs.txt", false);
 }