+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+### 2019.05.07
+- Creation of gameprefs text file to store settings of the game
+- Changed FileLogger to use strings for printing instead of const char*
+- Creation of FileReader class
+	- Implementation of getSettings function to retrieve the settings of the game stored in gameprefs file
+
+**Next:**
+- Open window using display settings specified in gameprefs.txt
+
+
 ### 2019.04.07
 - Implemented basic directory structure
 - Creation of FileLogger class to easily write to files and log messages